fonts: add last-resort
The glyphs of the Last Resort fonts are used as the backup of “last resort” to any other font: if a font cannot represent any particular Unicode character, the appropriate “missing” glyph from the Last Resort fonts is displayed instead. This provides users with the ability to more easily discern what type of character it is, and provides a clue as to what type of font they would need to display the characters properly. For more information, see The Unicode Standard, Section 5.3, Unknown and Missing Characters.